Euro teken bij price gedaan (95% done?)
Heb soufiane's code in mijn code verwerk. Het werkt (heb ook gechecked) enige issue = als je vluchten sorteer op datum de <date gaan weg uit de json? so gotta fix that 😵 en heb net alleen euro teken bij gedaan

+++ Main/Airline/DataModels/Flight.cs
@@ -7,10 +7,10 @@
     public DateTime BoardingDate;
     public DateTime EstimatedArrival;
     public Destination Destination;
-    public int MinPrice;
+    public double MinPrice;
 
 
-    public Flight(int flightNo, string flightid, string dayOrNight, Airplane airplane, DateTime boardingdate, DateTime estimatedarrival, Destination destination, int minPrice)
+    public Flight(int flightNo, string flightid, string dayOrNight, Airplane airplane, DateTime boardingdate, DateTime estimatedarrival, Destination destination, double minPrice)
     {
         FlightNo = flightNo; 
         FlightId = flightid;
